fix(sweep): empirical validation of curated axes — 5 defects found and fixed
User asked to validate that the curated ranges/steps do what we think
with a sweep of a subset. Built examples/sweep_validate.rs: encodes the
default stratum + every single-deviation stratum of modes_full (197
cells x 6 q-points) on 3 CID22-512 photos + noise/complex/checker
synthetics + one 64x64 tiny, ~32s. Hard-fails on inert steps,
fingerprint-contract violations, exact-trial contract violations, and
ordering breakage; soft direction checks per the provenance table;
per-label dsize/dssim2 aggregates; TSV committed to
benchmarks/sweep_validate_2026-06-10.tsv.

First run caught five real defects, all fixed:

1. Cell id collisions: lambda2/delta-DC/coupling-exponent probes
   rendered identically to their base configs. id() now renders every
   output-relevant trellis field that deviates from default; uniqueness
   enforced by new test across modes_full.
2. FINGERPRINT CONTRACT VIOLATION: TrellisSpeedMode was excluded as
   'output-neutral by construction' but it bounds the coefficient
   search on high-entropy blocks — Adaptive vs Thorough diverged by
   582 bytes on 512^2 noise at q95 (2 bytes on a photo). Now hashed
   (all four variants incl. Custom tiers); regression test added.
3. Curated coupling steps +-4 were UNCLAMPED, reproducing the
   historical screenshot-destruction mode: SSIM2 -31 (and -54 for the
   exponent probe) on noise at q85 while shedding up to 90% of bytes.
   trellis_coupled() now clamps lambda-adjustment to +-1.0; curated
   axes carry no unclamped coupling (enforced by test). Post-clamp the
   same cells score ~29 SSIM2.
4. ENTROPY_TRIAL_MAX_BYTES counterexample: CID22 1044329 q10 emits
   19.8KB progressive with baseline 2.0% smaller — above the 16KiB
   gate whose '7x margin' was calibrated on a smaller crossover. Gate
   raised to 32KiB; provenance rewritten with the counterexample.
5. SmallestSearch lost to ProgressiveMozjpeg by 0.09% on a CID22
   photo at q70: the scan search never trialed the canonical mozjpeg
   script. ScanStrategy::Search now trial-encodes it, making
   ProgressiveSearch/SmallestSearch byte-supersets of
   ProgressiveMozjpeg.

Also falsified two doc claims, corrected in place: trellis lambda does
NOT converge to no-trellis at high lambda (lambda=16 is ~15% LARGER and
+3 SSIM2 better than no-trellis — the ladder brackets it, monotone);
and restart markers are NOT strictly additive (RST resets re-base DC
prediction: a 4-row-interval baseline beat restart-free Smallest by 8
bytes once in 42 cells — kept as a documented WARN, Smallest stays
restart-free by design).

Validated as GOOD: all 33 single-deviation steps live (byte-diff rates
14%-100%, none inert); lambda ladder strictly monotone; all probe pairs
mutually distinct incl. [1,2] vs [2,1] chroma (Cb/Cr independently
wired); alias pairs byte-identical (auto_optimize/allow_16bit/Glassa);
direction signs match provenance (pre_blur shrinks, chroma 2x shrinks,
0.5x grows, cpl-4 < cpl+4); queue is defaults-first with non-decreasing
deviations; delta_dc=1.0 probe confirmed live but quality-collapsing at
q<=70 (-8..-36 SSIM2) — annotated as response-surface probe, not a
default candidate.

Collateral: cleared example lint debt so clippy --all-targets -D
warnings is green on BOTH default and full feature sets (migrated 6
examples off deprecated decode_jpeg_to_rgb — XYB-capable ones to
decode_jpeg_with_icc; zensim_regress ToleranceSpec/latest_preview API
drift; misc one-liners). VARIANT_GENERATION.md: pattern 6 'validate the
axes empirically', corrected fingerprint-exclusion guidance, byte-gate
and DC-reset lessons in known-limits.

Verified: sweep_validate ALL HARD CHECKS PASSED (1 expected WARN =
documented DC-reset anomaly); sweep unit tests 14/14; default suite
2159 passed / 0 failed; featured suite 2465 passed / 0 failed; clippy
-D warnings x2 green (default + featured, --all-targets).

+++ zenjpeg/examples/xyb_color_loss.rs
@@ -35,7 +35,7 @@ use zenjpeg::encode::encoder_types::{ChromaSubsampling, PixelLayout, XybSubsampl
 /// so DCT block boundaries don't smear adjacent grid samples together.
 fn build_grid_image(step: u32) -> (u32, u32, Vec<u8>, Vec<(u8, u8, u8)>) {
     assert!(step > 0 && step <= 256);
-    let n = (256 / step.max(1)) as u32; // samples per channel
+    let n = 256 / step.max(1); // samples per channel
     let total = n * n * n;
     let block = 8u32; // one 8×8 px block per sample so JPEG sees a uniform color
     let cols_per_row = 256u32; // wrap into a roughly square image
